ARM: UniPhier: add environment variable to update images in NAND
To boot UniPhier boards with the NAND boot mode, two images
(u-boot-spl.bin and u-boot-dtb.img) must be written at the correct
offset addresses.

TFTP downloading is useful to update such images in the NAND device.
We generally do:

  => nand erase 0 0x100000
  => tftpboot u-boot-spl.bin
  => nand write $loadaddr 0 0x10000
  => tftpboot u-boot-dtb.img
  => nand write $loadaddr 0x10000 0xf0000

It is a tedious and error-prone operation.

This commit provides the shorthand:

  => run nandupdate

Signed-off-by: Masahiro Yamada <yamada.m@jp.panasonic.com>