PD#143292: emmc: fix bootloader erase failure
A off&len check will be performed while erasing to protect keys.
when erasing BL, all copies on usr/boot0/boot1 will be ersaed.
For boot0&boot1, we will erase the whole partition as blk&cnt
are set as zero which will trigger this bug.

This bug was exposed since gxl revC as it's brom will try all 3
copies bl on emmc.

Change-Id: Ib679e64a15fe6780acecc147167b0ab353ce739c
Signed-off-by: Yonghui Yu <yonghui.yu@amlogic.com>