aes: add support of aes192 and aes256 [1/1]
PD#SWPL-131524

Problem:
1. There is no support for non-zero iv aes-cbc
2. There is no support for aes192 and aes256

Solution:
1. Cherry-pick patch from Андрей Мозжухин
aes: Allow non-zero initialization vector
AES encryption in CBC mode, in most cases, must be used with random
initialization vector. Using the same key and initialization vector several
times is weak and must be avoided.
Added iv parameter to the aes_cbc_encrypt_blocks and aes_cbc_decrypt_blocks
functions for passing initialization vector.
Command 'aes' now also require the initialization vector parameter.

2. Cherry-pick changes from Philippe Reynes from u-boot upstream
Commit messages from Philippe:
1. aes: add a define for the size of a block
In the code, we use the size of the key for the
size of the block. It's true when the key is 128 bits,
but it become false for key of 192 bits and 256 bits.
So to prepare the support of aes192  and 256,
we introduce a constant for the iaes block size.
2. aes: add support of aes192 and aes256
Until now, we only support aes128. This commit add the support
of aes192 and aes256.
3. aes: add test unit for aes128
This commit add test unit for aes128.
4. aes: add test unit for aes196 and aes256
This commit add test unit for aes196 and aes256.

Verify:
T5W

Change-Id: Iddccffca28f4d475f7b74f0f711b69ad5136ea2a
Signed-off-by: Andrey Mozzhuhin <amozzhuhin@yandex.ru>
Signed-off-by: Philippe Reynes <philippe.reynes@softathome.com>