gpt: repair the gpt table when head or backup gpt table is invalid
There is some gpt table errors when use open source tools to download
gpt table or is destroyed by other procedure.
Such as:
1.Get error storage size.
2.Download the gpt table in the error place.
3.The gpt table is destroyed by other procedure.

So test the gpt table is correct or not firstly. If the gpt table header is
error, update the correct gpt table. If the last partition size error in
the gpt table entry, get the size of current storage by dev_desc and update
the last partition size.

Change-Id: I0ef91a0f8462eca52924d17c5aaefcdae602267d
Signed-off-by: Jason Zhu <jason.zhu@rock-chips.com>