saradc: to initialize all of necessary registers [1/1]
PD#SWPL-5236

Problem:
some necessary registers are just initialized at the bl2 stage, the sampling
data of saradc will cause incorrent if the saradc code in bl2 is removed.

Solution:
in order to avoid the problem, to initialize all of necessary registers in
individual module.

Verify:
test pass on w400 and s400

Change-Id: I1433824117fc93d7b593263a3fbf0845d602a3aa
Signed-off-by: Xingyu Chen <xingyu.chen@amlogic.com>