env: access ENV with unified storage interface [1/1]
PD#SWPL-6469

Problem:
uboot 2018 can only use the default env which can not be changed and
saved so far.

Solution:
the amlogic storage is a interface which combines emmc, nand, nor, and
spi nand. using this interface, we can access these devices simply and
easily.
add the storage driver for env.

Verify:
verified on AD409 with spinor

Change-Id: I737f087ce099fe62b9df8b7b61ed4e1fbb6cf816
Signed-off-by: Sunny Luo <sunny.luo@amlogic.com>
Signed-off-by: Yonghui Yu <yonghui.yu@amlogic.com>