usb: usb event timeout cause panic [1/1]
PD#SWPL-6269

Problem:
Usb event timeout cause panic and the system restart. And
the reason of the problem is the sof period is inaccurate.

Solution:
Changing the register 0x54 to 0x2a, enabled the hs rx idle noise filter,
and this will improve reading and writing for u disk.
Clear the third bit of the GUSB2PHYCFGn and this will correct the sof period.

Verify:
g12b revB, PASS

Change-Id: I2d0221c08038ff59374f45a0c99912a250c517ff
Signed-off-by: he.he <he.he@amlogic.com>