gpt: return 1 directly when test the mbr sector
In sd-update card, we use RKPARM partition in bootloader to load
firmware, and use MS-DOS partition in recovery to update system.
Now, we want to use gpt in bootloader and abandon the RKPARM
partition. So in new sd-update card, we write the MS-DOS partition
table and gpt to sd card. Then we must return 1 directly when test
the mbr sector otherwise the gpt is unavailable.

Change-Id: I994a0cb86e7829c6edadfe357cbf51ad00b65cec
Signed-off-by: Jason Zhu <jason.zhu@rock-chips.com>