common: android: support boot-fastboot
In android Q, the bootloader must implement a new Boot Control
Block (BCB) command of boot-fastboot in order to support fastbootd.
Then the bootloader is expected to load and boot into the recovery
image upon seeing boot-fastboot in the BCB command. Recovery then
parses the BCB message and switches to fastbootd mode.

Change-Id: Ie868912071f6a371ec83caa8d591281fc063e6ae
Signed-off-by: Jason Zhu <jason.zhu@rock-chips.com>