aes: add support of aes192 and aes256 [1/1]
PD#SWPL-131524

Problem:
There is no support for aes192 and aes256

Solution:
Cherry-pick changes from Philippe Reynes from u-boot upstream
Commit messages from Philippe:
1. aes: add a define for the size of a block
In the code, we use the size of the key for the
size of the block. It's true when the key is 128 bits,
but it become false for key of 192 bits and 256 bits.
So to prepare the support of aes192  and 256,
we introduce a constant for the iaes block size.
2. aes: add support of aes192 and aes256
Until now, we only support aes128. This commit add the support
of aes192 and aes256.
3. aes: add test unit for aes128
This commit add test unit for aes128.
4. aes: add test unit for aes196 and aes256
This commit add test unit for aes196 and aes256.

Verify:
S905C2

Change-Id: I6a952c8904b6f409cb06660db75142a4c26848ee
Signed-off-by: Philippe Reynes <philippe.reynes@softathome.com>
Reviewed-by: Simon Glass <sjg@chromium.org>