rockchip: rkimg: use menu key to enter bootrom download mode
It's possible that USB is disabled due to developer needs
a critial size of u-boot.bin.

Disabling USB makes vbus can't be detected any more, so that
we add menu key and the events trigger are changed:

- rockusb mode(fallback to bootrom mode):
	"recovery key pressed + vbus=1" => "menu key pressed"

- recovery mode:
	"recovery key pressed + vbus=0" => "recovery key pressed"

At the most time, USB is enabled and this feature is not applied.

Signed-off-by: Joseph Chen <chenjh@rock-chips.com>
Change-Id: I9da8f17987742cea3267dbb35ee88d57a00c578a